Backend of an investment platform
This is a backend of an investment platform where salary earners who cannot afford big projects like buying of houses, cars, starting of businesses etc., are made to pair with others to save, daily, monthly etc, and the money realized is given to one person in the groove so they can start their projects. And before the payout time reaches, the platform owner uses this money to invest somewhere else, such as loaning out to loan companies etc, and payout to the investors when it is time for them to cashout, so he pays these investors with an interest. So if for instance, there are 10 investors in a group with each contributing $500, if the platform owner pays them with an interest of 1%, then they'll each realize $5,0050 instead of $5,000. There are more investment options available on the platform, but this is the one I can easily explain here to help make the reader understand how the codes work.